requirements.txt
streamlit
langchain
langchain-openai
beautifulsoup4
chromadb
python-dotenv